# packages/observing-shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Observ.ing API",
    "version": "0.1.0-alpha",
    "description": "**⚠️ This API is unstable and subject to change without notice.**\n\nDecentralized biodiversity observation platform API built on AT Protocol."
  },
  "servers": [
    {
      "url": "https://biosky-appview-378013734384.us-central1.run.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "description": "Decentralized identifier"
          },
          "handle": {
            "type": "string",
            "description": "User handle"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL"
          }
        },
        "required": [
          "did"
        ]
      },
      "Observer": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Profile"
          },
          {
            "type": "object",
            "properties": {
              "role": {
                "type": "string",
                "enum": [
                  "owner",
                  "co-observer"
                ],
                "description": "Observer role"
              }
            },
            "required": [
              "role"
            ]
          }
        ]
      },
      "Subject": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0,
            "description": "Subject index within the observation"
          },
          "communityId": {
            "type": "string",
            "description": "Community-agreed taxon name"
          },
          "identificationCount": {
            "type": "integer",
            "description": "Number of identifications"
          }
        },
        "required": [
          "index",
          "identificationCount"
        ]
      },
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "description": "Decimal latitude"
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "description": "Decimal longitude"
          },
          "uncertaintyMeters": {
            "type": "number",
            "description": "Coordinate uncertainty in meters"
          },
          "continent": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "stateProvince": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "waterBody": {
            "type": "string"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "Occurrence": {
        "type": "object",
        "properties": {
          "uri": {
            "type": "string",
            "description": "AT Protocol URI"
          },
          "cid": {
            "type": "string",
            "description": "Content identifier"
          },
          "observer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Profile"
              },
              {
                "description": "Primary observer"
              }
            ]
          },
          "observers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Observer"
            },
            "description": "All observers including co-observers"
          },
          "scientificName": {
            "type": "string",
            "description": "[DEPRECATED] Scientific name from record - use effectiveTaxonomy instead"
          },
          "communityId": {
            "type": "string",
            "description": "Community ID for subject 0 (backward compat)"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subject"
            },
            "description": "All subjects with their community IDs"
          },
          "eventDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date of observation"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "verbatimLocality": {
            "type": "string"
          },
          "occurrenceRemarks": {
            "type": "string"
          },
          "taxonId": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "taxonRank": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "vernacularName": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "kingdom": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "phylum": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "class": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "order": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "family": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "genus": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "effectiveTaxonomy": {
            "$ref": "#/components/schemas/EffectiveTaxonomy"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Image URLs"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "identificationCount": {
            "type": "integer"
          }
        },
        "required": [
          "uri",
          "cid",
          "observer",
          "observers",
          "subjects",
          "eventDate",
          "location",
          "images",
          "createdAt"
        ]
      },
      "EffectiveTaxonomy": {
        "type": "object",
        "properties": {
          "scientificName": {
            "type": "string"
          },
          "taxonId": {
            "type": "string"
          },
          "taxonRank": {
            "type": "string"
          },
          "vernacularName": {
            "type": "string"
          },
          "kingdom": {
            "type": "string"
          },
          "phylum": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "order": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "genus": {
            "type": "string"
          }
        },
        "required": [
          "scientificName"
        ],
        "description": "Taxonomy from winning identification"
      },
...
      "FeedFilters": {
        "type": "object",
        "properties": {
          "taxon": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "radius": {
            "type": "number"
          }
        }
      },
      "FeedResponse": {
        "type": "object",
        "properties": {
          "occurrences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Occurrence"
            }
          },
          "cursor": {
            "type": "string",
            "description": "Cursor for pagination"
          }
        },
        "required": [
          "occurrences"
        ]
      },
      "ExploreFeedResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FeedResponse"
          },
          {
            "type": "object",
            "properties": {
              "meta": {
                "type": "object",
                "properties": {
                  "filters": {
                    "$ref": "#/components/schemas/FeedFilters"
                  }
                }
              }
            }
          }
        ]
      },
      "HomeFeedResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FeedResponse"
          },
          {
            "type": "object",
            "properties": {
              "meta": {
                "type": "object",
                "properties": {
                  "followedCount": {
                    "type": "integer"
                  },
                  "nearbyCount": {
                    "type": "integer"
                  },
                  "totalFollows": {
                    "type": "integer"
                  }
                },
                "required": [
                  "followedCount",
                  "nearbyCount",
                  "totalFollows"
                ]
              }
            }
          }
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/feeds/explore": {
      "get": {
        "tags": [
          "Feeds"
        ],
        "summary": "Get explore feed with optional filters",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "taxon",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "lat",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "lng",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "radius",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Explore feed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploreFeedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeds/home": {
      "get": {
        "tags": [
          "Feeds"
        ],
        "summary": "Get personalized home feed (requires auth)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "lat",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "lng",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "nearbyRadius",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Home feed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeFeedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}